points earned work

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -27,7 +27,7 @@ class CreateUsersTable extends Migration
             $table->boolean('isInvestor');            
             $table->rememberToken();
 
-            $table->integer('distance_travelled')->default(0);
+            $table->unsignedBigInteger('distance_travelled')->default(0);
             $table->float('points', 9, 2)->default(0);
         });
     }
